refactor: extract runtime capability boundary
Restore the checkpoint resume exception dependency while moving the execution-gate state machine into its own module.
